test(runner): refactor config fixtures
